Added new sql queries to GIT

+++ R/raid_tables/multifactor_risk_blobtest_junctiontable.sql
@@ -1,12 +1,11 @@
 USE [FileTableDB]
 GO
 
-ALTER TABLE [dbo].[tMultiFactorRiskBlobTest] DROP CONSTRAINT [FK_FileTableDB_ModelTypeID_Betas_ModelType]
+
+IF EXISTS (SELECT * FROM sysobjects WHERE name = 'tMultiFactorRiskBlobTest' AND type = 'U')
+		DROP TABLE [dbo].[tMultiFactorRiskBlobTest]
 GO
 
-/****** Object:  Table [dbo].[tMultiFactorRiskBlobTest]    Script Date: 31/08/2016 14:14:05 ******/
-DROP TABLE [dbo].[tMultiFactorRiskBlobTest]
-GO
 
 /****** Object:  Table [dbo].[tMultiFactorRiskBlobTest]    Script Date: 31/08/2016 14:14:05 ******/
 SET ANSI_NULLS ON
@@ -19,17 +18,13 @@ SET ANSI_PADDING ON
 GO
 
 CREATE TABLE [dbo].[tMultiFactorRiskBlobTest](
-	[lTraderID] [bigint] NOT NULL,
-	[dtStartDate] [datetime] NOT NULL,
-	[dtEndDate] [datetime] NOT NULL,
-	[hPathLocator] [hierarchyid] NOT NULL,
-	[dtCreatedDate] [datetime] NULL,
-	[sCreatedByUserID] [varchar](50) NULL,
- CONSTRAINT [AK_FileTableDB_BlobTest_TraderIDStartEndCreatedUserFactor] UNIQUE NONCLUSTERED 
+	[bHashID] varbinary(64)  NOT NULL,
+	[hPathLocator] hierarchyid NOT NULL,
+	[dtCreatedDate] datetime NOT NULL,
+	[sCreatedByUserID] varchar(50) NOT NULL,
+ CONSTRAINT [AK_FileTableDB_BlobTest_HashIDCreatedDateCreatedBy] UNIQUE NONCLUSTERED 
 (
-	[lTraderID] ASC,
-	[dtStartDate] ASC,
-	[dtEndDate] ASC,
+	[bHashID] ASC,
 	[dtCreatedDate] ASC,
 	[sCreatedByUserID] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
